feat(agent-info): move version selector to header (above tabs)
Lifts version state from AgentDefinition up to AgentInfo and renders the
version dropdown in DetailPageLayout's selector slot, matching the pattern
used by API detail pages.
